Cleanup release routes
